diagram data: add badge stuff

+++ marketplace/marketplace-model/src/main/java/at/jku/cis/iVolunteer/model/diagram/xnet/data/XDiagramDataPoint.java
@@ -11,7 +11,7 @@ public class XDiagramDataPoint extends XDiagramData {
         this.value = value;
     }
 
-    public double getValue() {
+    public float getValue() {
         return this.value;
     }
